+++ R/vapour_input_geometry.R
@@ -13,12 +13,15 @@
 #' Note that `limit_n` and `skip_n` interact with the affect of `sql`, first the query is executed on the data source, then
 #' while looping through available features `skip_n` features are ignored, and then a feature-count begins and the loop
 #' is stopped if `limit_n` is reached.
+#'
+#' Note that `extent` applies to the 'SpatialFilter' of 'ExecuteSQL': https://www.gdal.org/ogr_sql.html#ogr_sql_exec_sql.
 #' @param dsource data source name (path to file, connection string, URL)
 #' @param layer integer of layer to work with, defaults to the first (0) or the name of the layer
 #' @param sql if not empty this is executed against the data source (layer will be ignored)
 #' @param textformat indicate text output format, available are "json" (default), "gml", "kml", "wkt"
 #' @param limit_n an arbitrary limit to the number of features scanned
 #' @param skip_n an arbitrary number of features to skip
+#' @param extent apply an arbitrary extent, only when 'sql' used (must be 'ex = c(xmin, xmax, ymin, ymax)')
 #' @examples
 #' file <- "list_locality_postcode_meander_valley.tab"
 #' ## A MapInfo TAB file with polygons
@@ -45,28 +48,44 @@
 #' @export
 #' @aliases vapour_read_geometry_text vapour_read_extent
 #' @name vapour_read_geometry
-vapour_read_geometry <- function(dsource, layer = 0L, sql = "", limit_n = NULL, skip_n = 0) {
+vapour_read_geometry <- function(dsource, layer = 0L, sql = "", limit_n = NULL, skip_n = 0, extent = NA) {
   if (!is.numeric(layer)) layer <- index_layer(dsource, layer)
   limit_n <- validate_limit_n(limit_n)
-  vapour_read_geometry_cpp(dsource = dsource, layer = layer, sql = sql, what = "geometry", textformat = "", limit_n = limit_n, skip_n = skip_n)
+  if (length(extent) > 1) {
+    if (!length(extent) == 4) stop("'extent' must be length 4 'c(xmin, xmax, ymin, ymax)'")
+  }
+
+  if (is.na(extent[1])) extent = 0.0
+
+  vapour_read_geometry_cpp(dsource = dsource, layer = layer, sql = sql, what = "geometry", textformat = "", limit_n = limit_n, skip_n = skip_n, ex = extent)
 }
 
 #' @export
 #' @rdname vapour_read_geometry
-vapour_read_geometry_text <- function(dsource, layer = 0L, sql = "", textformat = "json", limit_n = NULL, skip_n = 0) {
+vapour_read_geometry_text <- function(dsource, layer = 0L, sql = "", textformat = "json", limit_n = NULL, skip_n = 0, extent = NA) {
   if (!is.numeric(layer)) layer <- index_layer(dsource, layer)
   textformat = match.arg (tolower (textformat), c ("json", "gml", "kml", "wkt"))
   limit_n <- validate_limit_n(limit_n)
-  vapour_read_geometry_cpp(dsource = dsource, layer = layer, sql = sql, what = "text", textformat = textformat, limit_n = limit_n, skip_n = skip_n)
+  if (length(extent) > 1) {
+    if (!length(extent) == 4) stop("'extent' must be length 4 'c(xmin, xmax, ymin, ymax)'")
+  }
+  if (is.na(extent[1])) extent = 0.0
+
+  vapour_read_geometry_cpp(dsource = dsource, layer = layer, sql = sql, what = "text", textformat = textformat, limit_n = limit_n, skip_n = skip_n, ex = extent)
 }
 
 
 #' @rdname vapour_read_geometry
 #' @export
-vapour_read_extent <- function(dsource, layer = 0L, sql = "", limit_n = NULL, skip_n = 0) {
+vapour_read_extent <- function(dsource, layer = 0L, sql = "", limit_n = NULL, skip_n = 0, extent = NA) {
   if (!is.numeric(layer)) layer <- index_layer(dsource, layer)
   limit_n <- validate_limit_n(limit_n)
-  out <- vapour_read_geometry_cpp(dsource = dsource, layer = layer, sql = sql, what = "extent", textformat = "", limit_n = limit_n, , skip_n = skip_n)
+  if (length(extent) > 1) {
+    if (!length(extent) == 4) stop("'extent' must be length 4 'c(xmin, xmax, ymin, ymax)'")
+  }
+  if (is.na(extent[1])) extent = 0.0
+
+  out <- vapour_read_geometry_cpp(dsource = dsource, layer = layer, sql = sql, what = "extent", textformat = "", limit_n = limit_n, , skip_n = skip_n, ex = extent)
   nulls <- unlist(lapply(out, is.null))
   if (any(nulls)) out[nulls] <- replicate(sum(nulls), rep(NA_real_, 4L), simplify = FALSE)
   out