+++ tests/testthat/test-utils.R
@@ -1,0 +1,5 @@
+context("test-utils")
+
+test_that("utilities work", {
+  expect_equal(2 * 2, 4)
+})